updated JwtTokenRepository and authentication service

+++ src/main/kotlin/com/nlc/forensic/repository/JwtTokenRepository.kt
@@ -8,7 +8,7 @@ import java.util.*
 interface JwtTokenRepository: JpaRepository<JwtToken, Int> {
 
     @Query("SELECT t FROM JwtToken t JOIN t.user u WHERE u.id = :userId AND t.loggedOut = false")
-    fun findAllTokensByUser(userId: Int): List<JwtToken>
+    fun findAllTokensByUser(userId: Long): List<JwtToken>
 
     fun findByToken(token: String): Optional<JwtToken>
 }
